refactor: extract SSE utility tests to separate file to reduce complexity
- Split routes.test.ts (613 lines) by extracting SSE utility function tests
- Create dedicated routes.sse.test.ts for SSE sanitization and writeSSEData tests
- Reduce main routes.test.ts to 608 lines, addressing Codacy complexity warning
- Maintain all test coverage with 38 route tests + 13 SSE utility tests
- All 51 tests continue to pass after refactoring
